+++ liblzma-rs/src/common/stream_flags_decoder.rs
@@ -18,11 +18,11 @@ extern "C" fn read32le(buf: *const u8) -> u32 {
 pub const LZMA_STREAM_FLAGS_SIZE: c_int = 2;
 extern "C" fn stream_flags_decode(options: *mut lzma_stream_flags, in_0: *const u8) -> bool {
     return unsafe {
-        if *in_0.offset(0) != 0 || *in_0.offset(1) as c_int & 0xf0 != 0 {
+        if *in_0.offset(0) != 0 || *in_0.offset(1) & 0xf0 != 0 {
             return true;
         }
         (*options).version = 0;
-        (*options).check = (*in_0.offset(1) as c_int & 0xf) as lzma_check;
+        (*options).check = (*in_0.offset(1) & 0xf) as lzma_check;
         false
     };
 }
