+++ liblzma-rs/src/lzma/lzma_encoder_optimum_fast.rs
@@ -156,8 +156,8 @@ pub unsafe extern "C" fn lzma_lzma_optimum_fast(
     let mut i: u32 = 0;
     while i < REPS as u32 {
         let buf_back: *const u8 = buf.offset(-((*coder).reps[i as usize] as isize)).offset(-1);
-        if !(*buf.offset(0) as c_int != *buf_back.offset(0) as c_int
-            || *buf.offset(1) as c_int != *buf_back.offset(1) as c_int)
+        if !(*buf.offset(0) != *buf_back.offset(0)
+            || *buf.offset(1) != *buf_back.offset(1))
         {
             let len: u32 = lzma_memcmplen(buf, buf_back, 2, buf_avail) as u32;
             if len >= nice_len {
