+++ liblzma-rs/src/simple/powerpc.rs
@@ -41,7 +41,7 @@ unsafe extern "C" fn powerpc_code(
             *buffer.offset(i.wrapping_add(1) as isize) = (dest >> 16) as u8;
             *buffer.offset(i.wrapping_add(2) as isize) = (dest >> 8) as u8;
             let ref mut fresh0 = *buffer.offset(i.wrapping_add(3) as isize);
-            *fresh0 = (*fresh0 as c_int & 0x3) as u8;
+            *fresh0 = (*fresh0 & 0x3) as u8;
             let ref mut fresh1 = *buffer.offset(i.wrapping_add(3) as isize);
             *fresh1 = (*fresh1 as u32 | dest) as u8;
         }
